+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Certaines zones ne sont pas encore peuplées correctement, ou les lettres de cib
 
 ## Fix :
 - [#568](https://github.com/Refhi/Weda-Helper/issues/568) - la date de fin d'une alerte spécifiée dans les options de Weda-Helper n'est plus obligatoire.
+- [#570](https://github.com/Refhi/Weda-Helper/issues/570) - la recherche des motifs d'arrêt de travail utilise désormais la liste officielle utilisée sur ameli.fr
+- [#574](https://github.com/Refhi/Weda-Helper/issues/574) - amélioration du message d'erreur si SumatraPDF n'est pas installé
 
 # [2.15.1.2] - fixes et améliorations
 ## L’historique des facturations dans la page des FSE est désormais ouverte !